+++ zetaclient/config/config_chain.go
@@ -32,7 +32,9 @@ func New(setDefaults bool) Config {
 // bitcoinConfigRegnet contains Bitcoin config for regnet
 func bitcoinConfigRegnet() BTCConfig {
 	return BTCConfig{
-		RPCUsername:     "smoketest", // smoketest is the previous name for E2E test, we keep this name for compatibility between client versions in upgrade test
+		// `smoketest` is the previous name for E2E test,
+		// we keep this name for compatibility between client versions in upgrade test
+		RPCUsername:     "smoketest",
 		RPCPassword:     "123",
 		RPCHost:         "bitcoin:18443",
 		RPCParams:       "regtest",
@@ -59,30 +61,7 @@ func tonConfigLocalnet() TONConfig {
 // it contains list of EVM chains with empty endpoint except for localnet
 func evmChainsConfigs() map[int64]EVMConfig {
 	return map[int64]EVMConfig{
-		chains.Ethereum.ChainId: {
-			Chain: chains.Ethereum,
-		},
-		chains.BscMainnet.ChainId: {
-			Chain: chains.BscMainnet,
-		},
-		chains.Goerli.ChainId: {
-			Chain:    chains.Goerli,
-			Endpoint: "",
-		},
-		chains.Sepolia.ChainId: {
-			Chain:    chains.Sepolia,
-			Endpoint: "",
-		},
-		chains.BscTestnet.ChainId: {
-			Chain:    chains.BscTestnet,
-			Endpoint: "",
-		},
-		chains.Mumbai.ChainId: {
-			Chain:    chains.Mumbai,
-			Endpoint: "",
-		},
 		chains.GoerliLocalnet.ChainId: {
-			Chain:           chains.GoerliLocalnet,
 			Endpoint:        "http://eth:8545",
 			RPCAlertLatency: 60,
 		},